integer arithmetic types: do not use unsigned for lengths
In C it is not normally a good idea to use an unsigned integer type
for integer values, even if they are known not ever to be zero (for
example, because they are lengths).  This is because C unsigned
arithmetic has unhelpful behaviour when the values would become
negative.

In particular, comparing signed and unsigned integers, and doing
arithmetic (especially subtraction) when unsigned integers are
present, can be dangerous and lead to unexpected results.

So fix the resulting warnings (which are due to -Wsign-compare which
comes from -W) by making all lengths, counts (and iterators over them)
and return values from scanf be of signed types, usually int32_t
instead of uint32_t (but occasionally int).

+++ log.c
@@ -19,7 +19,7 @@ static void vMessage(uint32_t class, const char *message, va_list args)
     FILE *dest=stdout;
 #define MESSAGE_BUFLEN 1023
     static char buff[MESSAGE_BUFLEN+1]={0,};
-    uint32_t bp;
+    size_t bp;
     char *nlp;
 
     if (secnet_is_daemon) {
